DSL Speed Tweaks
The changes in this commmit reduce the execution time for the DSL in Scenario 5 significantly in some scenarios with no impact on the end result.

Specifically, the second iteration of Scenario 6 went from around 430 minutes down to 6.

Some other non-functional changes are as follows and have no impact on the overall evaluation:

 - General testing framework impprovements
 - Disabling of old SanityTests that would sometimes cause test failures since they used an old "SimpleSolver" that is not compatible with consistent iterative changes